Crate automerge_protocol

Modules

error

Structs

ActorId

An actor id is a sequence of bytes. By default we use a uuid which can be nicely stack allocated.

Change

ChangeHash

CursorDiff

ListDiff

MapDiff

MultiElementInsert

ObjDiff

Op

OpId

Patch

RootDiff

A custom MapDiff that implicitly has the object_id Root and is a map object.

ScalarValues

We don’t implement Serialize/Deserialize b/c this struct will always be serialized as 2 fields that are part of a larger struct. (It will never be serialized as its own struct/map)

SortedVec

TableDiff

TextDiff

Enums

DataType

Diff

DiffEdit

ElementId

Key

MapType

ObjType

ObjectId

OpType

ParseChangeHashError

ScalarValue

ScalarValueKind

Auto-generated discriminant enum variants

SequenceType